Moved IRequiredFileResolver interface to where it's used
Avoids an 'import' in AbstractOpenDeclaration.
